pyproject.toml
[tool.poetry]
name = "search-vacancies"
version = "0.1.0"
description = ""
authors = ["Liudmila Beloritskaya <meiramirth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
numpy = "^2.0.0"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
xlrd = "^2.0.1"
yfinance = "^0.2.40"
accessify = "^0.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
flake8 = "^7.1.0"
mypy = "^1.10.0"
isort = "^5.13.2"
pytest-cov = "^5.0.0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'




[tool.poetry.group.lint.dependencies]
mypy = "^1.10.0"
black = "^24.4.2"
flake8 = "^7.0.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py312']
skip-string-normalization = false
exclude = """
/(
    \\.git
  | \\.mypy_cache
  | \\.tox
  | \\.venv
  | _build
  | buck-out
  | build
  | dist
)/
"""